Move header to __layout
