+++ src/routes/__layout.svelte
@@ -1,7 +1,9 @@
 <script lang="ts">
   import "../app.css";
+  import Header from "../lib/Header.svelte";
 </script>
 
+<Header />
 <main>
   <slot />
 </main>
