+++ src/routes/settings.svelte
@@ -1,0 +1,1 @@
+<h1>Hello world!</h1>
